requirements.txt
django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
httpx==0.26.0
pydantic==2.10.5
